services/transaction-service/requirements.txt
boto3==1.34.131
psycopg2-binary==2.9.9
prometheus-client==0.20.0
python-json-logger==2.0.7
